Change the use of flags in `compile_object' and `compile_if'
In function `compile_if': When compiling the predicate part of if
form, its value is expected and there are more expressions in the if
form.

In function `compile_object': Only compile a lambda form when its
return value would be used, and add a RETURN instruction when there's
no more expressions.