Fix a bug when replacing the label in some instructions
When call a function recursively multiple times, the label part in
instructions like JUMP, FJUMP, TJUMP and SAVE should not be replaced
more than twice. However, this is just a ugly hack for the source
code. The normal solution is to assemble the byte-code only once in
the compilation step.